refactor: change TestDiff to TestArchitectureDiff, and move Diffnode class in architecture.py to node.py
