+++ src/caret_analyze/architecture/architecture.py
@@ -32,6 +32,7 @@ from ..value_objects import (CallbackGroupStructValue, CallbackStructValue,
                              CommunicationStructValue, ExecutorStructValue,
                              NodePathStructValue, NodeStructValue, PathStructValue,
                              PublisherStructValue, ServiceStructValue, SubscriptionStructValue)
+from ..value_objects.node import DiffNode
 
 logger = logging.getLogger(__name__)
 
@@ -640,45 +641,3 @@ class DiffArchitecture:
         left_only_topics = tuple(set_left_topics - common_node_topics)
         right_only_topics = tuple(set_right_topics - common_node_topics)
         return left_only_topics, right_only_topics
-
-
-class DiffNode:
-
-    def __init__(
-        self,
-        left_node: NodeStructValue,
-        right_node: NodeStructValue
-    ):
-        self._left_node = left_node
-        self._right_node = right_node
-
-    def diff_node_pubs(self) -> Tuple[Tuple[str, ...], Tuple[str, ...]]:
-        set_left_pubs = set(self._left_node.publish_topic_names)
-        set_right_pubs = set(self._right_node.publish_topic_names)
-        common_node_pubs = set_left_pubs & set_right_pubs
-        left_only_pubs = tuple(set_left_pubs - common_node_pubs)
-        right_only_pubs = tuple(set_right_pubs - common_node_pubs)
-        return left_only_pubs, right_only_pubs
-
-    def diff_node_subs(self) -> Tuple[Tuple[str, ...], Tuple[str, ...]]:
-        set_left_subs = set(self._left_node.subscribe_topic_names)
-        set_right_subs = set(self._right_node.subscribe_topic_names)
-        common_node_subs = set_left_subs & set_right_subs
-        left_only_subs = tuple(set_left_subs - common_node_subs)
-        right_only_subs = tuple(set_right_subs - common_node_subs)
-        return left_only_subs, right_only_subs
-
-    def diff_node_callbacks(self) -> Tuple[Tuple[str, ...], Tuple[str, ...]]:
-        if self._left_node.callback_names is None:
-            set_left_callbacks = set()
-        else:
-            set_left_callbacks = set(self._left_node.callback_names)
-        if self._right_node.callback_names is None:
-            set_right_callbacks = set()
-        else:
-            set_right_callbacks = set(self._right_node.callback_names)
-
-        common_node_callbacks = set_left_callbacks & set_right_callbacks
-        left_only_callbacks = tuple(set_left_callbacks - common_node_callbacks)
-        right_only_callbacks = tuple(set_right_callbacks - common_node_callbacks)
-        return left_only_callbacks, right_only_callbacks
