+++ src/caret_analyze/value_objects/node.py
@@ -270,3 +270,45 @@ class NodeStructValue(ValueObject, Summarizable):
         d['paths'] = [_.summary for _ in self.paths]
 
         return d
+
+
+class DiffNode:
+
+    def __init__(
+        self,
+        left_node: NodeStructValue,
+        right_node: NodeStructValue
+    ):
+        self._left_node = left_node
+        self._right_node = right_node
+
+    def diff_node_pubs(self) -> Tuple[Tuple[str, ...], Tuple[str, ...]]:
+        set_left_pubs = set(self._left_node.publish_topic_names)
+        set_right_pubs = set(self._right_node.publish_topic_names)
+        common_node_pubs = set_left_pubs & set_right_pubs
+        left_only_pubs = tuple(set_left_pubs - common_node_pubs)
+        right_only_pubs = tuple(set_right_pubs - common_node_pubs)
+        return left_only_pubs, right_only_pubs
+
+    def diff_node_subs(self) -> Tuple[Tuple[str, ...], Tuple[str, ...]]:
+        set_left_subs = set(self._left_node.subscribe_topic_names)
+        set_right_subs = set(self._right_node.subscribe_topic_names)
+        common_node_subs = set_left_subs & set_right_subs
+        left_only_subs = tuple(set_left_subs - common_node_subs)
+        right_only_subs = tuple(set_right_subs - common_node_subs)
+        return left_only_subs, right_only_subs
+
+    def diff_node_callbacks(self) -> Tuple[Tuple[str, ...], Tuple[str, ...]]:
+        if self._left_node.callback_names is None:
+            set_left_callbacks = set()
+        else:
+            set_left_callbacks = set(self._left_node.callback_names)
+        if self._right_node.callback_names is None:
+            set_right_callbacks = set()
+        else:
+            set_right_callbacks = set(self._right_node.callback_names)
+
+        common_node_callbacks = set_left_callbacks & set_right_callbacks
+        left_only_callbacks = tuple(set_left_callbacks - common_node_callbacks)
+        right_only_callbacks = tuple(set_right_callbacks - common_node_callbacks)
+        return left_only_callbacks, right_only_callbacks
